+++ src/Part1/BOJ10869/B10869_jjeonghak.java
@@ -1,25 +1,29 @@
-package Part1.BOJ10872;
+package Part1.BOJ10869;
 
 import java.io.*;
 
-public class Pt1_10872_김정한 {
+public class B10869_jjeonghak {
 
 	public static void main(String[] args) throws IOException {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 		StringBuilder sb = new StringBuilder();
 
-		int n = Integer.parseInt(br.readLine());
-		long ans = 1;
+		String[] input = br.readLine().split(" ");
+		int a = Integer.parseInt(input[0]);
+		int b = Integer.parseInt(input[1]);
 
-		while (n != 0) {
-			ans *= n--;
-		}
+		sb
+			.append(a + b).append("\n")
+			.append(a - b).append("\n")
+			.append(a * b).append("\n")
+			.append(a / b).append("\n")
+			.append(a % b).append("\n");
 
-		sb.append(ans);
 		bw.write(sb.toString());
 		bw.close();
 		br.close();
 	}
 
 }
+
