+++ src/Part1/BOJ10872/B10872_jjeonghak.java
@@ -1,16 +1,22 @@
-package Part1.BOJ11654;
+package Part1.BOJ10872;
 
 import java.io.*;
 
-public class Pt1_11654_김정한 {
+public class B10872_jjeonghak {
 
 	public static void main(String[] args) throws IOException {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 		StringBuilder sb = new StringBuilder();
 
-		sb.append((int)(br.readLine().charAt(0)));
+		int n = Integer.parseInt(br.readLine());
+		long ans = 1;
 
+		while (n != 0) {
+			ans *= n--;
+		}
+
+		sb.append(ans);
 		bw.write(sb.toString());
 		bw.close();
 		br.close();
