+++ src/Part1/BOJ2557/B2557_jjeonghak.java
@@ -1,6 +1,6 @@
 package Part1.BOJ2557;
 
-public class Pt1_2557_김정한 {
+public class B2557_jjeonghak {
 
 	public static void main(String[] args) {
 		System.out.println("Hello World!");
