+++ src/Part1/BOJ2753/B2753_jjeonghak.java
@@ -1,27 +1,23 @@
-package Part1.BOJ2588;
+package Part1.BOJ2753;
 
 import java.io.*;
 
-public class Pt1_2588_김정한 {
+public class B2753_jjeonghak {
 
 	public static void main(String[] args) throws IOException {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
 		StringBuilder sb = new StringBuilder();
 
-		int a = Integer.parseInt(br.readLine());
-		int b = Integer.parseInt(br.readLine());
-
-		int iter = b;
-		while (iter != 0) {
-			sb.append(a * (iter % 10)).append("\n");
-			iter /= 10;
-		}
-		sb.append(a * b).append("\n");
+		sb.append(check(Integer.parseInt(br.readLine())));
 
 		bw.write(sb.toString());
 		bw.close();
 		br.close();
+	}
+
+	private static int check(int n) {
+		return (n % 4 == 0 && n % 100 != 0) || (n % 400 == 0) ? 1 : 0;
 	}
 
 }
